modesetting: Add DRI3 support to modesetting driver with glamor disabled
This patch enables OpenGL/ES applications to use the HW accelerated
graphics driver while the Xserver composition uses the software rendered.

This new addition to the modesetting driver can be enabled adding
the msdri3 parameter on the AccelMethod option. So glamor will not
be enabled for the KMS modesetting driver.

Section "Device"
       Identifier "kms"
       Driver "modesetting"
       Option "AccelMethod" "msdri3"
EndSection

The motivation of this work is to avoid Xserver to use CMA memory
on Raspberry Pi [123] VC4 driver, so Xserver doesn't suffer on
allocation memory errors that would crash desktop environment when
we run out of GPU/CMA memory.

This setup is only expected to be used for pre-RPi4 hw, as RPI4 V3D has
MMU so the limitation of CMA availability is not an issue for the
Xserver regular operation.

Signed-off-by: Christopher Michael <cmichael@igalia.com>
Signed-off-by: Jose Maria Casanova Crespo <jmcasanova@igalia.com>